bit structer

+++ app/src/main/java/RecipApp/Ingredient.java
@@ -28,7 +28,7 @@ public class Ingredient {
     name = newName;
   }
 
-  public void editPrice(int newPrice) {
+  public void editPrice(Double newPrice) {
     price = newPrice;
   }
